Improve ProductCard.
